# apps/api/artifacts/swagger/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_welcome","summary":"App welcome","parameters":[],"responses":{"200":{"description":"Welcome message"}},"tags":["Welcome"]}},"/auth/login":{"post":{"operationId":"AuthPostController_login","summary":"Login user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserRequest"}}}},"responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserResponse"}}}},"401":{"description":"Unauthorized"}},"security":[{"ApiKey":[]}],"tags":["Auth"]}},"/auth/refresh-token":{"post":{"operationId":"AuthPostController_refreshToken","summary":"Refresh user token","parameters":[],"responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenResponse"}}}},"400":{"description":"Bad request"}},"security":[{"bearer":[]}],"tags":["Auth"]}},"/users":{"post":{"operationId":"UserPostController_run","summary":"Create user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserResponse"}}}},"400":{"description":"Bad Request"}},"security":[{"ApiKey":[]}],"tags":["Users"]}},"/messages":{"get":{"operationId":"MessageGetController_findByCriteria","summary":"Find messages by criteria","parameters":[{"name":"eventName","required":false,"in":"query","example":"user.created","description":"Message eventName","schema":{"type":"string"}},{"name":"sortColumn","required":false,"in":"query","description":"Sort the response by class property","example":"name","schema":{"type":"string"}},{"name":"sortOrder","required":false,"in":"query","description":"Sort the response by ASC or DESC","example":"ASC","schema":{"type":"string"}},{"name":"take","required":false,"in":"query","description":"Number of records to obtain","example":2,"schema":{"type":"number"}},{"name":"page","required":false,"in":"query","description":"Page number of the records list","example":1,"schema":{"type":"number"}}],"responses":{"200":{"description":"Ok","headers":{"X-RateLimit-Limit: 100":{"description":"The max calls by minute limit control","type":"number"},"X-RateLimit-Remaining : 99":{"description":"Remaining calls in this time period","type":"number"},"X-RateLimit-Reset: 1556970499":{"description":"Time until next RateLimit reset","type":"number"},"page:1":{"description":"Current page of the record set","type":"number"},"count:100":{"description":"Total count of records","type":"number"},"limit:10":{"description":"Max number of records returned","type":"number"}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindMessagesByCriteriaResponse"}}}},"404":{"description":"Bad request"}},"security":[{"bearer":[]}],"tags":["Messages"]}}},"info":{"title":"DDD Skeleton","description":"API boilerplate from DDD","version":"v1","contact":{"name":"Cristian Sánchez","url":"","email":"devcristiansanchez@gmail.com"}},"tags":[],"servers":[{"url":"/v1/","description":"v1"}],"components":{"securitySchemes":{"ApiKey":{"type":"apiKey","in":"header","name":"Authorization"},"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginUserRequest":{"type":"object","properties":{"email":{"type":"string","example":"desarrollo@mail.es"},"password":{"type":"string","example":"strongPassword"}},"required":["email","password"]},"LoginUserResponse":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjYzkxNzIyYS1lN2VkLTRkMjgtYWZhYi05OGNjMzhjZDY4NTEiLCJhdWQiOlsiR0VORVJBTCJdLCJpYXQiOjE2NzU1MjY1OTIsImV4cCI6MjMwNjY3ODU5Mn0.587F5hFKrKPpHs4wNSA0FD-JLuous2KEzE5kpzFxRfM"},"refreshToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjYzkxNzIyYS1lN2VkLTRkMjgtYWZhYi05OGNjMzhjZDY4NTEiLCJhdWQiOlsiR0VORVJBTCJdLCJpYXQiOjE2NzU1MjY1OTIsImV4cCI6MTY3NTUyNzE5Mn0.Rco1kYLa2Z02jclmIRvc2n8gjcjYrla_cuS6sG8drII"}},"required":["accessToken","refreshToken"]},"RefreshTokenResponse":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjYzkxNzIyYS1lN2VkLTRkMjgtYWZhYi05OGNjMzhjZDY4NTEiLCJhdWQiOlsiR0VORVJBTCJdLCJpYXQiOjE2NzU1MjY1OTIsImV4cCI6MjMwNjY3ODU5Mn0.587F5hFKrKPpHs4wNSA0FD-JLuous2KEzE5kpzFxRfM"},"refreshToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjYzkxNzIyYS1lN2VkLTRkMjgtYWZhYi05OGNjMzhjZDY4NTEiLCJhdWQiOlsiR0VORVJBTCJdLCJpYXQiOjE2NzU1MjY1OTIsImV4cCI6MTY3NTUyNzE5Mn0.Rco1kYLa2Z02jclmIRvc2n8gjcjYrla_cuS6sG8drII"}},"required":["accessToken","refreshToken"]},"CreateUserRequest":{"type":"object","properties":{"name":{"type":"string","example":"API"},"email":{"type":"string","example":"desarrollo@mail.es"},"password":{"type":"string","example":"strongPassword"}},"required":["name","email","password"]},"CreateUserResponse":{"type":"object","properties":{"id":{"type":"string","example":"61f4722b-cfd2-48e8-bd8f-7b8af7c5f285"},"name":{"type":"string","example":"API"},"email":{"type":"string","example":"desarrollo@mail.es"}},"required":["id","name","email"]},"FindMessagesByCriteriaResponse":{"type":"object","properties":{"data":{"example":[{"id":"d594db68-ae8b-425c-a519-ab4bce45b07a","eventId":"d8fdcd22-4c8f-4fa7-9a5f-257519c9b313","eventName":"user.created","data":{"id":"01e88d1f-ffbe-40a1-a14b-b2f8a24a7447","name":"John Doe","email":"john.doe@mail.com","password":"*"}},{"id":"f9dd4f99-321e-483f-84ab-1f6bc61fff61","eventId":"cf2ca631-0624-4a08-af6e-2114444d6b31","eventName":"user.created","data":{"id":"2503c4cf-dda0-4c41-a556-3f69de8fa0f1","name":"Jane Doe","email":"jane.doe@mail.com","password":"*"}}],"type":"array","items":{"type":"string"}}},"required":["data"]}}}}